requirements.txt
alembic==1.13.1
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozendict==2.4.4
html5lib==1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
Mako==1.3.5
MarkupSafe==2.1.5
multitasking==0.0.11
numpy==1.26.4
pandas==2.2.2
pandas-ta==0.3.14b0
peewee==3.17.5
platformdirs==4.2.2
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scipy==1.14.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.3
yfinance==0.2.40
